--- db_structures.php
@@ -1,15 +1,0 @@
-<?php
-function data()
-{
-    $db = new PDO(
-        'mysql:host=DB;dbname=collectorapp',
-        'root',
-        'password'
-
-    );
-    return $db;
-}
-
-$db = data();
-
-
