+++ styles.css
@@ -1,9 +1,12 @@
 * {
     box-sizing: border-box;
     margin: 10px;
+
 }
 
+
 .collection_item {
+
     width: 150px;
     height: 150px;
     display: flex;
@@ -11,10 +14,15 @@
     background-color: antiquewhite;
     padding: 10px;
     margin: 10px;
+
+
 }
 
 .item_container{
-    width: 100%;
-    height: 20%;
-    display: flex
+    width: 1500px;
+    height: 200px;
+    display: flex;
+    overflow-y: auto;
+    overflow-x: auto;
+
 }
